Backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
django-otp==1.2.0
qrcode==7.4.2
numpy==1.24.3
opencv-python==4.8.1.78
scikit-learn==1.3.2
tensorflow==2.13.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
mysqlclient==2.2.0
gunicorn==21.2.0
whitenoise==6.6.0
